requirements.txt
weaviate-client==4.6.3
torch>=2.0.0
transformers>=4.35.0
datasets>=2.15.0
tqdm>=4.66.0